[UPDATE VERSION 0.6.1 PYTHON] - Added comments, Added a ValueError to capture disconnections from Arduino, Added feature logging oncoming messages along with correcting minor mistake of functions being called I2C.
